Move ExtData.rc offline lightning and ocean mask entries to met text files
Since ocean mask and offline lightning are both meteorology-dependent
they are now moved from the ExtData.rc template files to the meteorology
text files used during run directory creation. This new design has the
added benefit of allowing a different met-source for offline lightning for
testing of GEOS-IT which does not yet have lightning files available, and
for allowing a non-standard file path for GEOS-IT ocean mask on the NASA
discover cluster.

Note that entries in ExtData.rc are not necessary required to run a
simulation. It is therefore okay to include ocean mask in ExtData.rc for
all smulations eventhough it is only used for transport tracers.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>