Updates GCHP run directory creation for GEOS-FP and GEOS-IT
These updates streamline the choosing of GEOS-FP and GEOS-IT
files, either native or processed, on discover or not, and
whether to use mass fluxes or not. The mass flux and derived
winds option is still there in GEOS-FP but is not included
in the options during GEOS-IT run directory creation. Users can
change to using mass fluxes at run-time within the directory.
Winds are default.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>